Proposal launcher: done
