+++ package.json
@@ -25,7 +25,6 @@
     "numeral": "^2.0.6",
     "react": "^16.13.1",
     "react-apollo": "^3.1.5",
-    "react-bootstrap": "^1.4.0",
     "react-dnd": "^5.0.0",
     "react-dnd-html5-backend": "^3.0.2",
     "react-document-meta": "^3.0.0-beta.2",
@@ -40,7 +39,6 @@
     "react-switch": "^5.0.1",
     "react-tag-input": "^6.4.3",
     "react-tiny-link": "^3.6.0",
-    "rimble-ui": "^0.14.0",
     "web3": "^1.2.11",
     "web3modal": "^1.9.0",
     "xml-js": "^1.6.11"
