+++ src/components/ProposalLauncher/SharesRequested.jsx
