+++ src/components/ProposalLauncher/TributeOffered.jsx
